requirements.txt
fastapi
uvicorn[standard]
httpx
pytest
pytest-asyncio
asgi-lifespan
python-dotenv
